requirements.txt
pydantic
PyYAML
cx_Oracle
python-ldap
